Extract core init capability emissions
